+++ src/main/java/com/main/fashionstore/Controller/admin/MainController.java
@@ -1,14 +1,14 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("contact")
-public class  ContactController {
+@RequestMapping("admin")
+public class MainController {
     @GetMapping("")
     public String index() {
-        return "user/contact";
+        return "admin/index";
     }
 }
