+++ src/main/java/com/main/fashionstore/Controller/user/CartController.java
@@ -1,13 +1,14 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
 @Controller
-@RequestMapping("orderDetail")
-public class OrderDetailController {
+@RequestMapping("cart")
+public class CartController {
     @GetMapping("")
     public String index() {
-        return "user/orderDetail";
+        return "user/shoping-cart";
     }
 }
