+++ src/main/java/com/main/fashionstore/Controller/user/ContactController.java
@@ -1,15 +1,14 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-
-@RequestMapping("create")
-public class CreateAccountController {
+@RequestMapping("contact")
+public class  ContactController {
     @GetMapping("")
-    public String loginForm() {
-        return "user/create-account";
+    public String index() {
+        return "user/contact";
     }
 }
