+++ src/main/java/com/main/fashionstore/Controller/user/CreateAccountController.java
@@ -1,13 +1,15 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
 @Controller
-@RequestMapping("index")
-public class IndexController {
+
+@RequestMapping("create")
+public class CreateAccountController {
     @GetMapping("")
-    public String index() {
-        return "user/index";
+    public String loginForm() {
+        return "user/create-account";
     }
 }
