+++ src/main/java/com/main/fashionstore/Controller/user/IndexController.java
@@ -1,14 +1,13 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 @Controller
-@RequestMapping("account")
-public class AccountUserController {
+@RequestMapping("index")
+public class IndexController {
     @GetMapping("")
     public String index() {
-        return "user/account";
+        return "user/index";
     }
 }
