+++ src/main/java/com/main/fashionstore/Controller/user/OrderDetailController.java
@@ -1,14 +1,13 @@
-package com.main.fashionstore.controller.admin;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 @Controller
-@RequestMapping("admin")
-public class MainController {
+@RequestMapping("orderDetail")
+public class OrderDetailController {
     @GetMapping("")
     public String index() {
-        return "admin/index";
+        return "user/orderDetail";
     }
 }
