+++ src/main/java/com/main/fashionstore/Controller/user/ProductController.java
@@ -1,9 +1,8 @@
-package com.main.fashionstore.controller.user;
+package com.main.fashionstore.Controller.user;
 
 import org.springframework.web.bind.annotation.GetMapping;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping("product")
