--- src/main/java/com/main/fashionstore/controller/user/test.java
@@ -1,4 +1,0 @@
-package com.main.fashionstore.controller.user;
-
-public class test {
-}
